Add bulk district creation with token validation
Introduces a new storeAll method in DistrictController to allow bulk district creation for all provinces and their cities, protected by a token. Updates the province show view to include a modal for this action, and adds the corresponding route. Also updates the cities table migration to use unsignedBigInteger for province_id.

+++ database/migrations/2025_04_17_094024_create_cities_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
             $table->uuid('slug');
             $table->string('code');
             $table->string('name');
-            $table->unsignedInteger('province_id');
+            $table->unsignedBigInteger('province_id');
             $table->string('url')->nullable();
             $table->timestamps();
